+++ Consequences/Hazards/IDepthVelocityHazard.cs
@@ -2,5 +2,5 @@ namespace Consequences.Hazards;
 
 public interface IDepthVelocityHazard : IDepthHazard
 {
-    double Velocity { get; }
+    float Velocity { get; }
 }
